Move view ops into a separate directory
Summary: Grouping view ops into its own separate directory can make it
explicit that they are a special category.
