+++ lazy_tensor_core/lazy_tensor_core/csrc/ops/index_ops.cpp
@@ -14,10 +14,10 @@
 #include "lazy_tensor_core/csrc/ops/index_get.h"
 #include "lazy_tensor_core/csrc/ops/index_put.h"
 #include "lazy_tensor_core/csrc/ops/ops.h"
-#include "lazy_tensor_core/csrc/ops/permute.h"
 #include "lazy_tensor_core/csrc/ops/scalar.h"
 #include "lazy_tensor_core/csrc/tensor_aten_ops.h"
 #include "lazy_tensor_core/csrc/ts_backend/ts_shape_inference.h"
+#include "lazy_tensor_core/csrc/view_ops/permute.h"
 #include "lazy_tensors/computation_client/util.h"
 
 namespace torch_lazy_tensors {
