+++ lazy_tensor_core/lazy_tensor_core/csrc/view_ops/generic_slice.cpp
@@ -1,8 +1,7 @@
-#include "lazy_tensor_core/csrc/ops/generic_slice.h"
+#include "lazy_tensor_core/csrc/view_ops/generic_slice.h"
 
-#include "lazy_tensor_core/csrc/ts_backend/ts_shape_inference.h"
 #include "lazy_tensor_core/csrc/ops/ltc_ops.h"
-#include "lazy_tensors/computation_client/util.h"
+#include "lazy_tensor_core/csrc/ts_backend/ts_shape_inference.h"
 
 namespace torch_lazy_tensors {
 namespace ir {
@@ -16,8 +15,7 @@ GenericSlice::GenericSlice(const torch::lazy::Value& input,
                           torch::lazy::MHash(base_indices, sizes)),
       base_indices_(base_indices.begin(), base_indices.end()),
       sizes_(sizes.begin(), sizes.end()) {
-  SetShapeDeferred(
-      [&]() { return compiler::InferShape(this); });
+  SetShapeDeferred([&]() { return compiler::InferShape(this); });
 }
 
 std::string GenericSlice::ToString() const {
