+++ lazy_tensor_core/lazy_tensor_core/csrc/view_ops/permute.cpp
@@ -1,8 +1,7 @@
-#include "lazy_tensor_core/csrc/ops/permute.h"
+#include "lazy_tensor_core/csrc/view_ops/permute.h"
 
-#include "lazy_tensor_core/csrc/ts_backend/ts_shape_inference.h"
 #include "lazy_tensor_core/csrc/helpers.h"
-#include "lazy_tensors/computation_client/util.h"
+#include "lazy_tensor_core/csrc/ts_backend/ts_shape_inference.h"
 
 namespace torch_lazy_tensors {
 namespace ir {
@@ -12,8 +11,7 @@ Permute::Permute(const torch::lazy::Value& input, std::vector<int64_t> dims)
     : torch::lazy::TsNode(torch::lazy::OpKind(at::aten::permute), {input},
                           /*num_outputs=*/1, torch::lazy::MHash(dims)),
       dims_(std::move(dims)) {
-  SetShapeDeferred(
-      [&]() { return compiler::InferShape(this); });
+  SetShapeDeferred([&]() { return compiler::InferShape(this); });
 }
 
 std::string Permute::ToString() const {
